Enhance README.md and video processing logic for improved functionality
This commit updates the README.md to reflect new features, including advanced 3-head cropping logic and cut detection capabilities. It also modifies the command line options for smoother processing, changing the `--smooth-duration` parameter to accept seconds instead of frames. Additionally, the video processing code is refactored to convert the smooth duration from seconds to frames, ensuring consistent behavior across different processors. These changes improve user guidance and enhance the overall functionality of the video processing tool.

+++ src/cli.rs
@@ -31,9 +31,9 @@ pub struct Args {
     #[argh(option, default = "10.0")]
     pub smooth_percentage: f32,
 
-    /// smooth duration in frames
-    #[argh(option, default = "45")]
-    pub smooth_duration: usize,
+    /// smooth duration in seconds
+    #[argh(option, default = "1.5")]
+    pub smooth_duration: f32,
 
     /// object probability threshold
     #[argh(option, default = "0.7")]
